+++ src/api/user.js
@@ -15,14 +15,10 @@ export function createUser (app, models, checkAuth, checkAdmin) {
     email: {
       isEmail: true,
       errorMessage: 'Invalid Email'
-    },
-    role: {
-      isString: true,
-      errorMessage: 'Invalid Role'
     }
   }), async (req, res) => {
     try {
-      const { name, email, role } = req.body
+      const { name, email } = req.body
 
       const errors = validationResult(req)
 
@@ -47,7 +43,7 @@ export function createUser (app, models, checkAuth, checkAdmin) {
         where: {
           name,
           email,
-          role
+          role: 'user'
         },
         defaults: {
           joined: new Date()
@@ -71,7 +67,7 @@ export function deleteUser (app, models, checkAuth, checkAdmin) {
     }
   }), async (req, res) => {
     try {
-      const { email } = req.query
+      const { email } = req.body
 
       const errors = validationResult(req)
 
@@ -89,6 +85,10 @@ export function deleteUser (app, models, checkAuth, checkAdmin) {
         return res.status(404).send('User does not exists')
       }
 
+      if (user.get('role') === 'admin') {
+        return res.status(403).send('You can not delete admins')
+      }
+
       const rowDeleted = await models.User.destroy({
         where: {
           email
@@ -100,7 +100,26 @@ export function deleteUser (app, models, checkAuth, checkAdmin) {
       }
       return res.status(200).send('Deleted User')
     } catch (error) {
-      return res.send(500).send('')
+      console.log(error)
+      return res.status(500).send('')
+    }
+  })
+}
+
+export function getUsers (app, models, checkAuth, checkAdmin) {
+  app.get('/api/users', checkAuth, checkAdmin, async (req, res) => {
+    try {
+      const users = await models.User.findAll({
+        attributes: [ 'id', 'name', 'email', 'role' ]
+      })
+
+      if (!users) {
+        return res.status(500).send('Users not found')
+      }
+
+      return res.status(200).send(users)
+    } catch (error) {
+      return res.status(500).send('')
     }
   })
 }
